Split the drawing tests to avoid Chrome crashes
