+++ tests/e2e/support/commands.js
@@ -4,6 +4,8 @@
 // https://on.cypress.io/custom-commands
 // ***********************************************
 
+const { MapBrowserEvent } = require('ol')
+
 const API_SERVICE_ICON_BASE_URL = 'https://service-icons.bgdi-dev.swisstopo.cloud/'
 
 const addIconsetsFixtureAndIntercept = () => {
@@ -140,3 +142,99 @@ const mockupBackendResponse = (endpoint, wantedOutput, aliasName) => {
     }).as(aliasName)
 }
 Cypress.Commands.add('mockupBackendResponse', mockupBackendResponse)
+
+Cypress.Commands.add('getDrawingMap', () => {
+    return cy.readWindowValue('drawingMap')
+})
+
+Cypress.Commands.add('clickDrawingTool', (name) => {
+    expect(['marker', 'text', 'line', 'measure']).to.include(name)
+    cy.get(`[data-cy="drawing-${name}`).click()
+    cy.readStoreValue('state.drawing.mode').should('eq', name.toUpperCase())
+})
+
+Cypress.Commands.add('readDrawingFeatures', (type, callback) => {
+    cy.readWindowValue('drawingManager')
+        .then((manager) => manager.source.getFeatures())
+        .then((features) => {
+            expect(features).to.have.length(1)
+            const foundType = features[0].getGeometry().getType()
+            expect(foundType).to.equal(type)
+            if (callback) callback(features)
+        })
+})
+
+Cypress.Commands.add('readDrawnGeoJSON', () => cy.readStoreValue('state.drawing.geoJson'))
+
+Cypress.Commands.add('checkDrawnGeoJsonProperty', (key, expected) => {
+    cy.readWindowValue('drawingManager')
+        .then((manager) => manager.source.getFeatures())
+        .then((features) => {
+            expect(features).to.have.length(1)
+            const v = features[0].get(key)
+            expect(v).to.equal(
+                expected,
+                `${v} != ${expected} Properties are ${JSON.stringify(
+                    features[0].getProperties(),
+                    null,
+                    2
+                )}`
+            )
+        })
+})
+
+/**
+ * This function has been taken from the OL draw spec. Simulates a browser event on the map
+ * viewport. The client x/y location will be adjusted as if the map were centered at 0,0.
+ *
+ * @param {string} type Event type.
+ * @param {number} x Horizontal offset from map center.
+ * @param {number} y Vertical offset from map center.
+ * @param {boolean} [opt_shiftKey] Shift key is pressed.
+ * @param {number} [opt_pointerId] Pointer id.
+ * @returns {MapBrowserEvent} The simulated event.
+ */
+function simulateEvent(map, type, x, y, opt_shiftKey, opt_pointerId = 0) {
+    cy.log(`simulating ${type} at [${x}, ${y}]`)
+
+    var viewport = map.getViewport()
+    let position = viewport.getBoundingClientRect()
+
+    // calculated in case body has top < 0 (test runner with small window)
+    const shiftKey = opt_shiftKey !== undefined ? opt_shiftKey : false
+    const event = {}
+    event.type = type
+    event.target = viewport.firstChild
+    event.clientX = position.left + x + position.width / 2
+    event.clientY = position.top + y + position.height / 2
+    event.shiftKey = shiftKey
+    event.preventDefault = function () {}
+    event.pointerType = 'mouse'
+    event.pointerId = opt_pointerId
+    event.isPrimary = true
+    event.button = 0
+    // @ts-ignore
+    const simulatedEvent = new MapBrowserEvent(type, map, event)
+    map.handleMapBrowserEvent(simulatedEvent)
+    return simulatedEvent
+}
+
+Cypress.Commands.add('clickDrawingMap', (x, y, callback) => {
+    cy.getDrawingMap().then((map) => {
+        // Create a point, a geojson will appear in the store
+        simulateEvent(map, 'pointermove', x, y)
+        simulateEvent(map, 'pointerdown', x, y)
+        simulateEvent(map, 'pointerup', x, y)
+        if (callback) callback(map)
+    })
+})
+
+Cypress.Commands.add('dragDrawingMap', (sx, sy, tx, ty, callback) => {
+    cy.getDrawingMap().then((map) => {
+        simulateEvent(map, 'pointerdown', sx, sy)
+        simulateEvent(map, 'pointermove', tx, ty)
+        simulateEvent(map, 'pointerdrag', tx, ty)
+        simulateEvent(map, 'pointerup', tx, ty)
+        if (callback) callback(map)
+    })
+})
